Update german.ini (#322)
I changed ü ö and ä to ue oe and ae, bc they weren't shown correctly (test it to see how it looks without and with the changes) and changed a few other things. Its WIP tho. With the help of a friend, I might even be able to create a norvegian version as well