Fixed backwards option scrolling (#380)
Fixed an issue where if you scrolled an option backwards, like the languages option, when going from option 0, the result of the modulus wasn't the latest option but -1. With this change, the size of the option list is added so that the result of the modulus is the same, and it doesn't have to deal with negative values.