Remove unused translation files [ENGLISH]
The reason why I'm removing them is because of the end user.
When someone would want to make/update a translation, they will base it off the english one. But then, like, what happens is that they translate more than they need to because of the old and unused strings being left in here.