Allow M3 DS Real to use YSMenu as flashcart loader (#1632)
Some notes to consider:

This will only work when also using the provided Autoboot, or by loading TWiLight from YSMenu directly.

I could not think of another way to avoid accidentally having any other cart using the "M3DS" DLDI name launch YSMenu. The DLDI name must match either the one used in TWiLight Menu++, or YSMenu.

Also, the original M3 DS DLDI does not work on TWiLight Menu++ anyway. The DLDI has a 32KB section and TWiLight Menu++ expects 16KB or less.

A workaround for this is to, well, not use the original M3 DS DLDI when launching TWiLight using the official kernel. This was already added to the Sakura kernel in DS-Homebrew/flashcard-archive, but obviously not everyone uses this. (but considering you cannot autoboot to the Sakura kernel, this point is moot. You're basically going from Sakura -> TWL -> YSMenu. Just use YSMenu or TWiLight at that point.)

As a result of the points above, I just decided to limit this to the M3 DS carts that support YSMenu. This was done by setting friendlyName to detect M3DS DLDI as opposed to M3DS.