Merge Ace3DS+ and R4iLS autoboot folders
This can reduce confusion for people who buy a flashcart that is a clone of one of these.
At the moment these are widely known to be Ace3DS+ clones, but it turns out they can also be clones of the R4iLS as well.
This is also done on DS-Homebrew/flashcard-archive, and has been for a while.

Fix #1689