booter_fc: work around the R4i-SDHC protection scheme
One of many of Taiju Yamada (nee Xenon++)'s tools included a small patcher
to work around the protection within the R4i-SDHC firmware.

As a result of this patcher, the entry address had to be changed to 0x02000450.
Fortunately, our Acekard 2 booter uses a 0x02000450 entry address, so it is
reused here.