Roll Acekard 2i, R4 Ultra and R4iTT flashcart loaders into a single loader
After reimplementing the AK2i DLDI driver, a theory was made that the BL2CK
loader, designed for the R4iTT, which is a descendant of the AK2i, would work
OOTB on its ancestors.

This indeed does work, and compatibility has been fine with most people on
r/flashcarts being happy with it. So these 3 loaders are all rolled into one
loader.