booter_fc: fix R4 Ultra (and co.) ARM9 entry address
Apparently newer versions of devkitARM fixed
an issue in the linker that caused the ARM9
entry address to be incorrectly configured.
And by doing so they broke the workaround
measure we used for the incorrect entry
address problem.

This does make something simpler: we can just
switch to the existing address_adapt ARM9 and
just set the entry address that way, and drop
the now-broken r4ig ARM9.