Update ESRB.ini (add online notice for WarioWare: D.I.Y.) (#2560)
* Update ESRB.ini

Add rating for Trace Memory

* Update ESRB.ini

Revised a few game titles, and added Game & Watch Collection 1 and 2.

* Update ESRB.ini

* Update ESRB.ini

* Update ESRB.ini (Remove M&S Beijing 2008 and Vancouver 2010)

These games show an ESRB rating on startup, so they need to be removed to show the ratings once.

* Update ESRB.ini (Remove M&S Beijing 2008 and Vancouver 2010)

These games show an ESRB rating on startup, so they need to be removed to show the ratings once.

* Update ESRB.ini (Remove M&S Beijing 2008 and Vancouver 2010)

These games show an ESRB rating on startup, so they need to be removed to show the ratings once.

* Add ESRB ratings for Deca Sports DS, Gardening Mama, WALL-E, and Wonder Pets! Save the Animals!

* Add ESRB ratings for Deca Sports DS, Gardening Mama, WALL-E, and Wonder Pets! Save the Animals!

* Add ESRB ratings for Deca Sports DS, Gardening Mama, WALL-E, and Wonder Pets! Save the Animals!

* Update ESRB.ini

Add rating for Disgaea DS

* Update ESRB.ini

Add rating for Disgaea DS

* Update ESRB.ini

Add rating for Disgaea DS

* Add online notice for WarioWare: D.I.Y.

* Add online notice for WarioWare: D.I.Y.

* Add online notice for WarioWare: D.I.Y.