Fix glitchy palettes for some animated icons
For some reason, palettes seem to update faster than the graphics when the next icon of the animation is drawn. This causes an issue where there is 1 frame in which the palette may be incorrect everytime the sequence updates to the next frame. As a hacky fix, I delay loading the new palette until the next vblank.